split generate commands
