+++ src/commands/redteam.ts
@@ -7,7 +7,7 @@ import * as path from 'path';
 import logger from '../logger';
 import { ALL_PLUGINS, DEFAULT_PLUGINS, subCategoryDescriptions } from '../redteam/constants';
 import telemetry from '../telemetry';
-import { doGenerateRedteam } from './generate';
+import { doGenerateRedteam } from './generate/redteam';
 
 interface RunRedteamOptions {
   config: string;
